More typescript
