+++ src/components/members-search/styles.tsx
@@ -1,8 +1,8 @@
 import { css, keyframes } from '@emotion/react'
+import styled from '@emotion/styled'
 import { Button } from 'hedvig-ui/button'
 import { Input } from 'hedvig-ui/input'
 import { Search as SearchBootstrapIcon } from 'react-bootstrap-icons'
-import styled from '@emotion/styled'
 import { Link } from 'react-router-dom'
 
 const fadeIn = (max: number) =>
