+++ src/routes/QuestionsPageRoute.tsx
@@ -1,6 +1,4 @@
-import QuestionsPage from 'containers/QuestionsPage'
+import { QuestionsPage } from 'containers/QuestionsPage'
 import React from 'react'
 
-const QuestionsPageRoute = () => <QuestionsPage />
-
-export default QuestionsPageRoute
+export const QuestionsPageRoute = () => <QuestionsPage />
