+++ src/routes/index.ts
@@ -3,7 +3,7 @@ import DashboardPageRoute from 'routes/DashboardPageRoute'
 import LoginPageRoute from 'routes/LoginPageRoute'
 import MembersPageRoute from 'routes/MembersPageRoute'
 import PrivateRoute from 'routes/PrivateRoute'
-import QuestionsPageRoute from 'routes/QuestionsPageRoute'
+import { QuestionsPageRoute } from 'routes/QuestionsPageRoute'
 import { ToolsPageRoute } from 'routes/ToolsPageRoute'
 
 export default {
